requirement.txt
streamlit==1.14.0
pandas==2.0.1
Pillow==9.2.0
numpy==1.24.2
torch==2.1.0
torchvision==0.15.0
opencv-python-headless==4.7.0.72
transformers==4.34.0